Exit if a mapnik style can't be created.
If there is a problem initializing a map style, this patch will cause
the mapnik backend to exit (with 2). This causes the complete
tirex-backend-manager process to exit with 2. The previous behaviour was
for it to constantly attempt to initialize the map forevery

A mapnik style can fail to load for many reasons, such as missing
database tables. This change means incorrectly set up servers can be
detected quicker.